+++ app/Http/Controllers/SongController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Album;
 use App\Models\Song;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
@@ -10,6 +11,7 @@ use Intervention\Image\ImageManager;
 
 class SongController extends Controller
 {
+
     public function index(Request $request){
 
         //validación de los datos del formulario de filtro
@@ -29,15 +31,34 @@ class SongController extends Controller
         return view('songs.index', compact('songs'));
     }
 
+    public function show(Song $song){
+        $artist = $song->artist;
+        return view('songs.show', compact('song', 'artist'));
+    }
+
     public function create(){
 
-        $albums = auth()->user()->artist->albums;
+        $user = auth()->user();
+
+        if($user->rol === 'admin' ){
+             $albums = Album::all();
+        }
+        else if($user->artist){
+            $albums = auth()->user()->artist->albums;
+        }
+        else
+            return redirect()->route("home");
 
         return view('songs.create', compact('albums'));
     }
 
     public function store(Request $request){
-        
+
+        $user = auth()->user();
+
+        if($user->rol !== 'admin' && $user->artist === null)
+            return redirect()->route("home");
+
         $request->validate([
             'title' => 'required|string|max:30',
             'song-cover-input' => 'nullable|image|mimes:jpeg,png,jpg,gif',
@@ -76,14 +97,31 @@ class SongController extends Controller
             $song->save();
         }
 
-        return redirect()->route("songs.index");
+        if($user->rol === 'admin' ){
+            return redirect()->route("songs.index");
+        }
+        else if($user->artist){
+            return redirect()->route("artist.dashboard");
+        }
+
+        
     }
 
     public function edit(Song $song){
+
+        $user = auth()->user();
+        if($user->rol !== 'admin' && $user->artist === null)
+            return redirect()->route("home");
+
         return view('songs.edit', compact('song'));
     }
 
     public function update(Request $request, Song $song){
+
+        $user = auth()->user();
+
+        if($user->rol !== 'admin' && $user->artist === null)
+            return redirect()->route("home");
 
         $request->validate([
             'title' => 'required|string|max:30',
@@ -118,12 +156,29 @@ class SongController extends Controller
 
         $song->update($request->all());
         $song->save();
-        return redirect()->route("songs.show", $song);
+
+        if($user->rol === 'admin' ){
+            return redirect()->route("songs.show", $song);
+        }
+        else if($user->artist){
+            return redirect()->route("artist.dashboard");
+        }
 
     }
 
     public function destroy(Song $song){
+        
+        $user = auth()->user();
+        if($user->rol !== 'admin' && $user->artist === null)
+            return redirect()->route("home");
+
         $song->delete();
-        return redirect()->route('songs.index');
+
+        if($user->rol === 'admin' ){
+            return redirect()->route('songs.index');
+        }
+        else if($user->artist){
+            return redirect()->route("artist.dashboard");
+        }  
     }
 }
